requirements.txt
torch
pillow
numpy
pandas
requests
python-dotenv
imagebind
google-generativeai
kdbai-client
